@@ -1,8 +1,7 @@
 import React from 'react';
-import ReactDOM from 'react-dom';
+import {createRoot} from 'react-dom/client';
 import './index.css';
 import App from './App';
-import reportWebVitals from './reportWebVitals';
 import { Provider } from 'react-redux';
 import { createStore,applyMiddleware,compose } from 'redux'
 import thunk from 'redux-thunk';
@@ -15,13 +14,15 @@ const store = createStore(
     composeEnhancers(applyMiddleware(thunk))
 )
 
-ReactDOM.render(
+const root = createRoot(document.getElementById('root'));
+root.render(
   <React.StrictMode>
     <Provider store={store}>
       <App />
     </Provider>
   </React.StrictMode>,
-  document.getElementById('root')
 );
 
-reportWebVitals();
+
+
+
