@@ -2,7 +2,7 @@
 function App() {
   return (
     <div className="App">
-     
+      <h1>Hello world</h1>
     </div>
   );
 }
